requirements.txt
agno
mistralai
haystack-ai
haystack-integrations
qdrant-haystack
python-dotenv
pydantic
